backend/requirements.txt
django==5.1.*
djangorestframework==3.15.*
pyjwt==2.10.*
pymysql==1.1.*
cryptography==44.*
gunicorn==23.*
python-dotenv==1.1.*
ruff==0.9.*
pytest==8.3.*
pytest-django==4.9.*
